requirements.txt
Django==1.4.2
PIL==1.1.7
django-storages==1.1.5
wsgiref==0.1.2
